requirements.txt
Flask==2.0.1
scapy==2.4.5
psycopg2-binary==2.9.3
Werkzeug==2.0.2